The second view is created for the candidate. Card and button components
