+++ 57blocks_recruitment/src/ComponentsCandidate/firstView/CardOpportunities.jsx
