+++ 57blocks_recruitment/src/ComponentsCandidate/firstView/JobOpportunities.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Box, Typography } from '@mui/material/'
 import CardOpportunities from './CardOpportunities'
-import icoUser from './img/icoData.svg'
-import icoWeb from './img/icoWeb.svg'
-import icoBlockchain from './img/icoBlockchain.svg'
-import icoTitle from './img/title.svg'
+import icoUser from '../img/icoData.svg'
+import icoWeb from '../img/icoWeb.svg'
+import icoBlockchain from '../img/icoBlockchain.svg'
+import icoTitle from '../img/title.svg'
 
 const productBlockchain = [
   {
